balanceManager introduced
Imitating the stockManager we have moved all the cash balance logic into
a type called balanceManager.

The balanceManager maintains two pieces of information. The 'current'
balance and the 'available' balance. The 'current' balance represents the
total amount of cash the user owns. The 'available' balance represents
the cash which is not tied up in open BUY orders. This is to prevent a
user from submitting more BUY orders than he/she can afford.

NB: The total value of an order or match is (price * amount).

The process for selling a stock:

When a sell order is matched (PARTIAL or FULl)
1: Add the total value of the match to both the 'current' and
   'available' balances

The process for buying a stock:

When a buy order is submitted
1: Ensure that the total value of the order is less than or equal to the
   user's 'available' balance
2: Deduct the total value of the order from the user's 'available'
   balance

When a buy order is matched (PARTIAL or FULL)
1: Deduct the total value of the order from the user's 'current'
   balance. Note that we have already reduced the 'available' balance
   when the buy was submitted.

When a buy order is cancelled
1: Add the total value of the order back to the user's 'available'
   balance.