Price is now uint64, rather than int64
This prevents us from ever accidentally introducing a negative price.
Previously we were using negative prices to represent matched buys.

The rationale behind this was that it would allow clients to easily
identify whether a match was for a buy or a sell by testing the sign of
the price. In practice we need to find the order that has been matched
anyway, so we can determine whether the matched order was a buy or a
sell from the originating order directly.