Moved svr/user.go into client/ package
We have split the file into three separate files data.go, managers.go
and user.go.

This change was made to create a simple and task specific package to
model and run the users. It could be argued that the user modelling code
and the client server (the code that was previously in this package)
should be separate. In the future these two components may be separated,
but right now we prefer the simplicity of a single package solution.

This is a first step and is currently very messy. The next steps to
develop this package are:

1: Re-think the approach to data we are taking here. Should the json
types be moved into a separate package. I can't see a clear answer here.
2: Precreate users. The UserMaker.Connect(...) should establish a
connection with an existing user, rather than creating a random new one.
The mechanism for creating a new user is not clear at this stage and
that development will be deferred.
3: Create a 'replay-mode' where logged messages are replayed through the
system. Since the users are pre-created this will always bring the users
back to a consistent state w.r.t the logged messages.
4: Produce a user snapshotting system. This would mean that we could
take the state of a set of users and store it. Then we won't need to
replay every message, just the set of new messages since the snapshot
was taken.