Clients are now created and then connected to
client.client now has a connecter channel. This allows the ClientServer
to pass in channels connecting that client with an external (e.g.
websocket) system. We currently create a new client by first passing a
NEW_USER message into the ClientServer and then passing in the
connection details.

Unfortunately this creates a small data race, where the connection
details message could be picked up before the NEW_USER message (because
they come on different channels). To deal with this hazard when we
receive a message on the connecter channel we check to see if a user for
the message exists. If there a user exists, the connect message is
forwarded on immediately, otherwise we store it. Each time we create a
new user we must check to see if there are any outstanding connect
messages for it.

Now using msg.Message from the web browser through the server. Removed
client.ClientMessage. msg.Message now has json marked variable names.