client.TraderMaker.Connect(...) method added
This method allows an external system (e.g. a network connection) to
connect to an existing trader. The existing client.TraderMaker.Make(...)
method now uses this method after it has sent the NEW_TRADER message to
the client server.

The reason we wish to be able to connect to pre-existing traders is to
allow traders with no external connection to be created and have their
state maintained as trades are processed. This allows us to easily
replay message logs on the client server and reproduce the state of
every trader reliably. It does, however, raise some questions about
scalability and robustness as each trader will be bound to a single
machine, or possibly replicated over a fix set. May need to revisit this
decision in the future.