Trader state sent when connection is established
When a connection is established with a trader, using the connectTo(...)
method, the trader sends its current state immediately on the responses
channel.

We also removed a latent race condition where we were sending a
*msg.Message object and then using it when sending the state on the
responses channel. Now we carefully wait until the state has been sent
before we send the message on the intoSvr channel. This is something to
look out for in the future and, at the appropriate time, we should
consider moving to passing around msg.Message values, instead of
pointers.