trader now detects when orders channel is closed
Previously when the orders channel was closed, for example when a
websocket connection is closed, the trader would start receiving nil
*msg.Message objects on the orders channel and panic with a nil
dereference. Now we detect this call the disconnect(...) method on
trader.

The disconnect(...) method on trader will close the existing responses
channel and nil both the responses and orders channel on trader.