Now replacing message 'kind' with a string value
The kind of a message, whether it is a SELL, BUY or CANCEL etc., is
encoded as an integer. When we display it on the web-page we substitute
the int with a readable string.

This now means we are maintaining the Kind type in two separate places.
We will need to keep an eye on this in case it becomes a pain.