Added first tests for the client package
A pair of very basic tests.

A test showing that the trader sends a specific disconnection message
when the orders channel is closed and then closes the responses channel.

A test showing that the trader sends a specific disconnection message
when a new connection message is received and that the trader then sends
a connection message to the new connection.

The utility method which starts the trader running also creates an
initial connection and tests that a specific connection message has been
sent.

In order to write these tests we had to add specific comments which are
sent for each connection and disconnection event.