Added unit tests for trader
Unit tests added:

Processing BUY, CANCEL, CANCELLED messages for an order.
Processing BUY, FULL messagese for an order, including the case where
the bid and match prices are different.
Processing BUY, PARTIAL messages for an order, including the case where
the bid and match prices are different.
Processing SELL, CANCEL, CANCELLED messages for an order.

As part of this we have produced a validate method for validating that a
trader is in a consistent state. A trader is in a consistent state iff:

1: The difference between the current balance and the available balance
(current - available) is the same as the total value of all outstanding
(not matched and not cancelled) BUY orders.

2: For every stock listed in 'stocks to sell' there exist outstanding
SELL orders summing to the quantity indicated, e.g. If 'stocks to sell'
indicates that the trader is selling 5 units of stock 1, then we expect
to see SELL orders for stock 1 whose quantity summs to 5.

We expect this validation criteria to expand in the future.