Added unit tests for trader
Processing SELL, FULL messagese for an order, including the case where
the offer and match prices are different.
Processing SELL, PARTIAL messages for an order, including the case where
the offer and match prices are different.

Also added a number of TODO comments about the code to record required
upcoming changes