response chan sending *client.Response not byte[]
Rather than marshalling the state of a trader into JSON as a byte[] we
are deep copying the trader's internal state and sending it in the
client.Response struct. This has several positive effects.

First we no longer need to have JSON friendly data-structures inside
trader. This allowed us to store stocks as a map[uint32]uint32 rather
than a map[string]uint32, removing a larget chunk of annoying strconv
code.

Second it makes the code much clearer. Previously the response was
returning a rather mysterious byte[] and you had to either read the code
that generated it, or rely on friendly comments to know what was
actually in that data.

Third and most importantly we allow different byte[] encodings to be
used without modifying the trader code.

This comes with the cost that we are constructing an intermediate
representation on every update of a trader.