Changed msg.Message.Amount/StockId to uint64
This means that every field in message, apart from Kind, TradeId and
TraderId, are uint64. This simplifies the code that deals with these
values. In particular this was motivated by the fact that the total
value of a BUY/SELL message is Amount * Price, previously Price was
uint64 and Amount uint32 making for awkward arithmetic.