trader's state now in single balanceManager struct
Previously we had three locations for the state of a trader. The
stockManager, which knew which stocks a trader owned and which one's
he/she was trying to sell. The balanceManager, which knew how much money
a trader had and how much was tied up in outstanding BUY orders. Finally
the trader itself contained a balanceManager and a stockManager knew
about all of the outstanding BUY/SELL/CANCEL orders for that trader as
well as the current trade-id.

These three components of state have been merged into a single
balanceManager struct. This provides two benefits.

1: The number of methods relating to processing messages is reduced.
Previously to process a FULL match for an oustanding BUY we had to
remove the matched BUY from the outstanding list, we had to complete the
BUY with the balanceManager and we had to complete the BUY with the
stocksManager. Now we can perform these actions in a single coherent
method.
2: It is easier to understand how the state of a trader changes with
each message without the extra indirection.