The inmemory coordinator now panics on io errors
This is because the inmemory coordinator expects perfect reads and
writes with some other in-memory component. If there is any source of
communication error it must be handled outside the read/write limits of
the coordinator.

For the current reference matching web-app this means that the code
inside the client/ package is expected to sanitise all messages coming
into the system.