Added Marshal and Unmarshal functions to msg/
The two functions define a binary format for Message structs. This
replaces the existing unsafe casting of structs to slices.

We have defined a 40 byte format using little endian byte ordering.
Neither of these decisions is very firm. In the future we will need to
take into consideration the possibility of data structures wrapping
Message structs being sent over the network. The reliable coordinator
is an example of this.

Casting a struct to a slice was unsuitable as a binary format because
it encodes the structure chosen by the compiler to represent that
struct. This could change between compiler version or across platforms.