Un/Marshal methods added to coordinator.RMessage
Previously we were serialising RMessage structs by casting the object to
an unsafe array and wrapping the array in a slice. This isn't
appropriate for a communicating between separate applications,
especially over a network.

We now expicitly serialise an RMessage as a 51 byte structure. The
wrapped msg.Message struct is serialised using its own Un/Marshal
methods.

We expect to add more meta-data to RMessage in the future, for instance
a checksum could be a good idea.