Broken RMessages dropped and not forwarded
Previously we would perform a best effort unmarshalling of an RMessage
coming into the reliable listener. If an error occurred the message
would still be forwarded on to the application and to the responder. Now
we simply log the error and drop the message on the floor.

This decision is based on two assumptions which do not yet appear in the
code base. One, that we log all incoming byte reads, so we can replay
the good messages and debug the bad ones. Two, that when the reliable
listener misses a message, because it was droppped or corrupted, it will
aggressively ask for it to be reset.

We have altered to mocked io.Reader found in treliable_test.go.
Now it takes a slice of bool to indicate a range of error/non-error
return values and a slice of int to indicate a range of read sizes. This
allows us to force the mocked io.Reader to send a bad message, which
will be dropped by the listener, and then send a good message so we can
verify the first message was processed and ignored.