bin/perf/ package now tests reads and writes
Previously we were only testing how fast we could write onto a channel.

Now we start a goroutine with the matching engine. Then we start a
goroutine sending order messages to the matching engine. Finally we
reuse the current goroutine to read messsages coming out of the matching
engine until a msg.SHUTDOWN message is received.

The timing is made from when we start reading messsages from the
matching engine until the msg.SHUTDOWN message is received. It is
possible for this to under report the actual running time, if the writer
and matching engine get a large head start on the reader.