Channels hidden behind MsgReader/Writer interfaces
Previously the coordinator package used, and exposed, channels of type
'chan *msg.Message'. These were used to pass messages into and out of
the matching engine.

Now we have created a pair of interfaces

MsgReader
	Read() *msg.Message
MsgWriter
	Write(*msg.Message)

These allow us to swap out channels for other means of sending messages.
A few useful implementations can be found in coordinator/readwrite.go.

A problem created by the use of interfaces is that we cannot combine
channels using 'select' blocks anymore. This meant that testing code for
matcher/tcoordinator_test.go had to be rewritten. But luckily this was a
simplification of the code.

Initial performance tests were disappointing and need further
investigation.

the bin/svr package has been broken by these changes and not fixed.